Partially rename program_* to script_*; Fix #14
Revision 99e88178 introduces a regression where the program_name variable triggers nounset exception due to the design decision of local variables can't mask/shadow a readonly global variable with the same name.  This patch fixes it, while improving the primitive variables again to the better UX script_*.

`program_basecommand` is not renamed as it is different in the runtime sense.

Refer-to: Variable with the same name can't be declared in the local scope if the global one is declared read-only <https://lists.gnu.org/archive/html/bug-bash/2020-11/msg00065.html>
Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>