Support preparation of a source image with over 4GiB filesize WIM image to a FAT32 filesystem
This patch implements the necessary logic to prepare a Windows
installation pendrive with a >4GiB filesize WIM archive in the
source image.

It uses the split functionality of the wimlib software to split oversized
WIM archives to be under 4GiB.  logic is
also implemented to ensure proper accounting of the prepare progress.

Signed-off-by: 林博仁(Buo-ren, Lin) <Buo.Ren.Lin@gmail.com>
Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>